requirements.txt
#### Paketi potrebni za program ####
deep-translator >= 1.5.4
IMDbPY >= 2021.4.18
requests >= 2.25.1
crg == 0.0.1
